+++ structure/animegan.go
@@ -1,0 +1,10 @@
+package structure
+
+type SendDataToPush struct {
+	Data   []string `json:"data"`
+	Action string   `json:"action"`
+}
+
+type SendDataStatus struct {
+	Hash string `json:"hash"`
+}
