+++ structure/bot.go
@@ -1,0 +1,12 @@
+package structure
+
+type EditDataToTlg struct {
+	Text      string `json:"text"`
+	ChatId    int    `json:"chat_id"`
+	MessageId int    `json:"message_id"`
+}
+
+type SendDataToTlg struct {
+	ChatId int    `json:"chat_id"`
+	Text   string `json:"text"`
+}
